Split server logic into shared modules for knip cleanup
- Extract shared bootstrap, open, keybinding, and receipt logic
- Move integration-only projection checkpoint repository support
- Update tests and web helpers to use the new shared modules
