+++ apps/server/src/bootstrap.shared.ts
@@ -1,0 +1,12 @@
+export function resolveFdPath(
+  fd: number,
+  platform: NodeJS.Platform = process.platform,
+): string | undefined {
+  if (platform === "linux") {
+    return `/proc/self/fd/${fd}`;
+  }
+  if (platform === "win32") {
+    return undefined;
+  }
+  return `/dev/fd/${fd}`;
+}
