+++ apps/server/src/main.logic.ts
@@ -1,0 +1,35 @@
+import { Effect } from "effect";
+
+import { ProjectionSnapshotQuery } from "./orchestration/Services/ProjectionSnapshotQuery";
+import { AnalyticsService } from "./telemetry/Services/AnalyticsService";
+
+export interface CliConfigShape {
+  readonly cwd: string;
+  readonly fixPath: Effect.Effect<void>;
+  readonly resolveStaticDir: Effect.Effect<string | undefined>;
+}
+
+export const recordStartupHeartbeat = Effect.gen(function* () {
+  const analytics = yield* AnalyticsService;
+  const projectionSnapshotQuery = yield* ProjectionSnapshotQuery;
+
+  const { threadCount, projectCount } = yield* projectionSnapshotQuery.getSnapshot().pipe(
+    Effect.map((snapshot) => ({
+      threadCount: snapshot.threads.length,
+      projectCount: snapshot.projects.length,
+    })),
+    Effect.catch((cause) =>
+      Effect.logWarning("failed to gather startup snapshot for telemetry", { cause }).pipe(
+        Effect.as({
+          threadCount: 0,
+          projectCount: 0,
+        }),
+      ),
+    ),
+  );
+
+  yield* analytics.record("server.boot.heartbeat", {
+    threadCount,
+    projectCount,
+  });
+});
