+++ apps/server/src/open.logic.ts
@@ -1,0 +1,112 @@
+import { spawn } from "node:child_process";
+
+import { EDITORS, type EditorId } from "@t3tools/contracts";
+import { Effect, Schema } from "effect";
+import { isCommandAvailable } from "./open.shared";
+
+export class OpenError extends Schema.TaggedErrorClass<OpenError>()("OpenError", {
+  message: Schema.String,
+  cause: Schema.optional(Schema.Defect),
+}) {}
+
+export interface OpenInEditorInput {
+  readonly cwd: string;
+  readonly editor: EditorId;
+}
+
+interface EditorLaunch {
+  readonly command: string;
+  readonly args: ReadonlyArray<string>;
+}
+
+const LINE_COLUMN_SUFFIX_PATTERN = /:\d+(?::\d+)?$/;
+
+function shouldUseGotoFlag(editor: (typeof EDITORS)[number], target: string): boolean {
+  return editor.supportsGoto && LINE_COLUMN_SUFFIX_PATTERN.test(target);
+}
+
+function fileManagerCommandForPlatform(platform: NodeJS.Platform): string {
+  switch (platform) {
+    case "darwin":
+      return "open";
+    case "win32":
+      return "explorer";
+    default:
+      return "xdg-open";
+  }
+}
+
+export function resolveAvailableEditors(
+  platform: NodeJS.Platform = process.platform,
+  env: NodeJS.ProcessEnv = process.env,
+): ReadonlyArray<EditorId> {
+  const available: EditorId[] = [];
+
+  for (const editor of EDITORS) {
+    const command = editor.command ?? fileManagerCommandForPlatform(platform);
+    if (isCommandAvailable(command, { platform, env })) {
+      available.push(editor.id);
+    }
+  }
+
+  return available;
+}
+
+export interface OpenShape {
+  readonly openBrowser: (target: string) => Effect.Effect<void, OpenError>;
+  readonly openInEditor: (input: OpenInEditorInput) => Effect.Effect<void, OpenError>;
+}
+
+export const resolveEditorLaunch = Effect.fnUntraced(function* (
+  input: OpenInEditorInput,
+  platform: NodeJS.Platform = process.platform,
+): Effect.fn.Return<EditorLaunch, OpenError> {
+  const editorDef = EDITORS.find((editor) => editor.id === input.editor);
+  if (!editorDef) {
+    return yield* new OpenError({ message: `Unknown editor: ${input.editor}` });
+  }
+
+  if (editorDef.command) {
+    return shouldUseGotoFlag(editorDef, input.cwd)
+      ? { command: editorDef.command, args: ["--goto", input.cwd] }
+      : { command: editorDef.command, args: [input.cwd] };
+  }
+
+  if (editorDef.id !== "file-manager") {
+    return yield* new OpenError({ message: `Unsupported editor: ${input.editor}` });
+  }
+
+  return { command: fileManagerCommandForPlatform(platform), args: [input.cwd] };
+});
+
+export const launchDetached = (launch: EditorLaunch) =>
+  Effect.gen(function* () {
+    if (!isCommandAvailable(launch.command)) {
+      return yield* new OpenError({ message: `Editor command not found: ${launch.command}` });
+    }
+
+    yield* Effect.callback<void, OpenError>((resume) => {
+      let child;
+      try {
+        child = spawn(launch.command, [...launch.args], {
+          detached: true,
+          stdio: "ignore",
+          shell: process.platform === "win32",
+        });
+      } catch (error) {
+        return resume(
+          Effect.fail(new OpenError({ message: "failed to spawn detached process", cause: error })),
+        );
+      }
+
+      const handleSpawn = () => {
+        child.unref();
+        resume(Effect.void);
+      };
+
+      child.once("spawn", handleSpawn);
+      child.once("error", (cause) =>
+        resume(Effect.fail(new OpenError({ message: "failed to spawn detached process", cause }))),
+      );
+    });
+  });
