+++ apps/server/src/orchestration/Services/RuntimeReceiptBus.shared.ts
@@ -1,0 +1,22 @@
+import { CheckpointRef, IsoDateTime, NonNegativeInt, ThreadId, TurnId } from "@t3tools/contracts";
+import { Schema } from "effect";
+
+export const CheckpointDiffFinalizedReceipt = Schema.Struct({
+  type: Schema.Literal("checkpoint.diff.finalized"),
+  threadId: ThreadId,
+  turnId: TurnId,
+  checkpointTurnCount: NonNegativeInt,
+  checkpointRef: CheckpointRef,
+  status: Schema.Literals(["ready", "missing", "error"]),
+  createdAt: IsoDateTime,
+});
+export type CheckpointDiffFinalizedReceipt = typeof CheckpointDiffFinalizedReceipt.Type;
+
+export const TurnProcessingQuiescedReceipt = Schema.Struct({
+  type: Schema.Literal("turn.processing.quiesced"),
+  threadId: ThreadId,
+  turnId: TurnId,
+  checkpointTurnCount: NonNegativeInt,
+  createdAt: IsoDateTime,
+});
+export type TurnProcessingQuiescedReceipt = typeof TurnProcessingQuiescedReceipt.Type;
