+++ apps/server/src/orchestration/Services/RuntimeReceiptBus.ts
@@ -1,6 +1,10 @@
-import { CheckpointRef, IsoDateTime, NonNegativeInt, ThreadId, TurnId } from "@t3tools/contracts";
+import { CheckpointRef, IsoDateTime, NonNegativeInt, ThreadId } from "@t3tools/contracts";
 import { Schema, ServiceMap } from "effect";
 import type { Effect, Stream } from "effect";
+import {
+  CheckpointDiffFinalizedReceipt,
+  TurnProcessingQuiescedReceipt,
+} from "./RuntimeReceiptBus.shared.ts";
 
 const CheckpointBaselineCapturedReceipt = Schema.Struct({
   type: Schema.Literal("checkpoint.baseline.captured"),
@@ -9,26 +13,6 @@ const CheckpointBaselineCapturedReceipt = Schema.Struct({
   checkpointRef: CheckpointRef,
   createdAt: IsoDateTime,
 });
-
-export const CheckpointDiffFinalizedReceipt = Schema.Struct({
-  type: Schema.Literal("checkpoint.diff.finalized"),
-  threadId: ThreadId,
-  turnId: TurnId,
-  checkpointTurnCount: NonNegativeInt,
-  checkpointRef: CheckpointRef,
-  status: Schema.Literals(["ready", "missing", "error"]),
-  createdAt: IsoDateTime,
-});
-export type CheckpointDiffFinalizedReceipt = typeof CheckpointDiffFinalizedReceipt.Type;
-
-export const TurnProcessingQuiescedReceipt = Schema.Struct({
-  type: Schema.Literal("turn.processing.quiesced"),
-  threadId: ThreadId,
-  turnId: TurnId,
-  checkpointTurnCount: NonNegativeInt,
-  createdAt: IsoDateTime,
-});
-export type TurnProcessingQuiescedReceipt = typeof TurnProcessingQuiescedReceipt.Type;
 
 export const OrchestrationRuntimeReceipt = Schema.Union([
   CheckpointBaselineCapturedReceipt,
