+++ apps/server/src/orchestration/commandInvariants.testing.ts
@@ -1,0 +1,41 @@
+export { findThreadById } from "./commandInvariants.shared.ts";
+
+import type {
+  OrchestrationCommand,
+  OrchestrationReadModel,
+  OrchestrationThread,
+  ProjectId,
+} from "@t3tools/contracts";
+import { Effect } from "effect";
+
+import { OrchestrationCommandInvariantError } from "./Errors.ts";
+
+function invariantError(commandType: string, detail: string): OrchestrationCommandInvariantError {
+  return new OrchestrationCommandInvariantError({
+    commandType,
+    detail,
+  });
+}
+
+export function listThreadsByProjectId(
+  readModel: OrchestrationReadModel,
+  projectId: ProjectId,
+): ReadonlyArray<OrchestrationThread> {
+  return readModel.threads.filter((thread) => thread.projectId === projectId);
+}
+
+export function requireNonNegativeInteger(input: {
+  readonly commandType: OrchestrationCommand["type"];
+  readonly field: string;
+  readonly value: number;
+}): Effect.Effect<void, OrchestrationCommandInvariantError> {
+  if (Number.isInteger(input.value) && input.value >= 0) {
+    return Effect.void;
+  }
+  return Effect.fail(
+    invariantError(
+      input.commandType,
+      `${input.field} must be an integer greater than or equal to 0.`,
+    ),
+  );
+}
