+++ apps/server/src/provider/Layers/ProviderRegistry.shared.ts
@@ -1,0 +1,7 @@
+import type { ServerProvider } from "@t3tools/contracts";
+import { Equal } from "effect";
+
+export const haveProvidersChanged = (
+  previousProviders: ReadonlyArray<ServerProvider>,
+  nextProviders: ReadonlyArray<ServerProvider>,
+): boolean => !Equal.equals(previousProviders, nextProviders);
