+++ apps/server/src/provider/Layers/ProviderRegistry.test.ts
@@ -30,7 +30,8 @@ import {
   readCodexConfigModelProvider,
 } from "./CodexProvider";
 import { checkClaudeProviderStatus, parseClaudeAuthStatusFromOutput } from "./ClaudeProvider";
-import { haveProvidersChanged, ProviderRegistryLive } from "./ProviderRegistry";
+import { ProviderRegistryLive } from "./ProviderRegistry";
+import { haveProvidersChanged } from "./ProviderRegistry.shared";
 import { ServerSettingsService, type ServerSettingsShape } from "../../serverSettings";
 import { ProviderRegistry } from "../Services/ProviderRegistry";
