+++ apps/web/src/components/chat/TraitsPicker.browser.tsx
@@ -17,12 +17,12 @@ import { render } from "vitest-browser-react";
 
 import { TraitsPicker } from "./TraitsPicker";
 import {
-  COMPOSER_DRAFT_STORAGE_KEY,
   ComposerThreadDraftState,
   useComposerDraftStore,
   useComposerThreadDraft,
   useEffectiveComposerModelState,
 } from "../../composerDraftStore";
+import { COMPOSER_DRAFT_STORAGE_KEY } from "../../composerDraftStore.shared";
 import { DEFAULT_CLIENT_SETTINGS } from "@t3tools/contracts/settings";
 
 // ── Claude TraitsPicker tests ─────────────────────────────────────────
