+++ apps/web/src/composerDraftStore.shared.ts
@@ -1,0 +1,1 @@
+export const COMPOSER_DRAFT_STORAGE_KEY = "t3code:composer-drafts:v1";
