+++ apps/web/src/hooks/useLocalStorage.shared.ts
@@ -1,0 +1,39 @@
+import * as Record from "effect/Record";
+import * as Schema from "effect/Schema";
+
+const isomorphicLocalStorage: Storage =
+  typeof window !== "undefined"
+    ? window.localStorage
+    : (function () {
+        const store = new Map<string, string>();
+        return {
+          clear: () => store.clear(),
+          getItem: (_) => store.get(_) ?? null,
+          key: (_) => Record.keys(store).at(_) ?? null,
+          get length() {
+            return store.size;
+          },
+          removeItem: (_) => store.delete(_),
+          setItem: (_, value) => store.set(_, value),
+        };
+      })();
+
+const decode = <T, E>(schema: Schema.Codec<T, E>, value: string) =>
+  Schema.decodeSync(Schema.fromJsonString(schema))(value);
+
+const encode = <T, E>(schema: Schema.Codec<T, E>, value: T) =>
+  Schema.encodeSync(Schema.fromJsonString(schema))(value);
+
+export const getLocalStorageItem = <T, E>(key: string, schema: Schema.Codec<T, E>): T | null => {
+  const item = isomorphicLocalStorage.getItem(key);
+  return item ? decode(schema, item) : null;
+};
+
+export const setLocalStorageItem = <T, E>(key: string, value: T, schema: Schema.Codec<T, E>) => {
+  const valueToSet = encode(schema, value);
+  isomorphicLocalStorage.setItem(key, valueToSet);
+};
+
+export const removeLocalStorageItem = (key: string) => {
+  isomorphicLocalStorage.removeItem(key);
+};
