+++ apps/web/src/lib/projectScriptKeybindings.ts
@@ -6,7 +6,7 @@ import {
 } from "@t3tools/contracts";
 import { Schema } from "effect";
 
-export const PROJECT_SCRIPT_KEYBINDING_INVALID_MESSAGE = "Invalid keybinding.";
+const PROJECT_SCRIPT_KEYBINDING_INVALID_MESSAGE = "Invalid keybinding.";
 
 function normalizeProjectScriptKeybindingInput(
   keybinding: string | null | undefined,
